Update linux/moz to use new perf format.
This is a test.  If this test passes, I'll convert the
rest of the measures.  It includes a change to the
expectations to account for this test's new performance.
If it didn't include that change, the line would be
improve: -9, regress: 9.

BUG=none
TEST=linux/moz is green

Review URL: http://codereview.chromium.org/3115012

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 994419a85f0e8657a9aa958c4990643ff6ecc3e9