Mac: Try to get tabpose unit tests working again.
The original problem was that the test plays an animation that outlives the test. When the animation is done, the tab strip model observer tries to unregister itself, but the registry died when the test exited already.

My faulty fix was to reset the observer as soon as the end animation starts. This has the problem that the end animation will try to access stale objects: TabContentss are destroyed when the test exits, but the exit animation still tries to access them when drawing ThumbnailLayers (because the observer didn't get a notification that these were dead, since it was unregistered).

This (hopefully better) fix now unregisters the observer directly before the tab strip goes away.

BUG= 54323,53893,54342
TEST=unit tests pass, valgrind stays green

Review URL: http://codereview.chromium.org/3304014

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f2d447cfe0e1383b664f73db319d8fd097b3e15f