Unit tests that use ProfileSyncServiceMock are leaking due to adding a DataSource (SyncResourcesSource) to ChromeURLDataManager.
The basic problem is that the destructors need to be called on the UI thread, but at shutdown the UI thread (which is the only thread left) is not recognized so the destruction is thrown to the thread even though the message loop is no longer running.  This means that the messages are dropped and memory is leaked.

I added cleanup code to the destructor.

BUG=70782
TEST=Valgrind

Review URL: http://codereview.chromium.org/6317017

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e9a6ff49bc07fdca1fa1107fe486ae8555c6a389