Fix valgrind error in -[TabStripController layoutTabsWithAnimation:regenerateSubviews]
The error was due to the fact that I was using the return values of calls to a nil object. This is only ok for 32bit values. For things like NSRects this means that most of the NSRect is not initialized.

Fix was to check for a nil object.

BUG=77884
TEST=Reproduced the error. Verified that this change fixed the error.

Review URL: http://codereview.chromium.org/6735047

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 43bd25e0f96c259a3dd4aeb0923ee62c236c7525