Fix leak in ProfileMenuButtonTest.MenuTest
Valgrind was reporting a leak in -[ProfileMenuDelegate menuWillOpen:].

I wasn't able to reproduce this on my machine with a Debug or Release build.

As a workaround I'm simply removing that part of the test. The test now just checks that popUpContextMenu:withEvent:forView: is called. This is still pretty good since it tests the mouse down code and the pressed state drawing code.

BUG=77910
TEST=Wasn't able to reproduce the same leak but I did get some other leaks in the menu code. Applied my patch and verified that valgrind ran with no errors.

Review URL: http://codereview.chromium.org/6677169

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e79d7dc73530396c27806a6794f4fcd65262926f