Add another newline before the TBR
Previously, the TBR notation appears on the first line of the description,
which prevented the commit-queue from recognizing it as a valid notation.  This
patch moves it to its own line, causing the patch to be accepted by the
commit-queue.

BUG=
TEST=


Review URL: http://codereview.chromium.org/7617013

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9a3d652cd700936140bceabc486a203693103fcd