Revert 100542 - Fix memory leak in sync notifier unit tests
When XmppConnection is deleted, its TaskPump member is not deleted immediately.
The destructor equeues a task to the message loop to have it deleted later.  A
comment in ~XmppConnection that explains why it does this.

In the unit tests, the TaskPump would get leaked because the message loop would
never get a chance to run that deletion task.  This commit fixes the problem
by running all pending tasks on that message loop before we exit the test.

BUG=80238,79651
TEST=InvalidationNotifierTest


Review URL: http://codereview.chromium.org/7863005

TBR=rlarocque@chromium.org
Review URL: http://codereview.chromium.org/7867024

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d18969dd541aa6db82ab302c08d091de167406e9