Fix a leak in MalwareDetailsTest.HTTPCache
The TestURLRequestContextGetter ensures that it is deleted by the IO thread.
If the IO thread doesn't exist when it is deleted, then the object is leaked.

This change stops the IO thread a bit later in the teardown, which ensure that
it is around long enough to handle TestURLRequestContextGetter's deletion.

It looks like willchan had fixed this error for a while with r85378.  At that
time, the profile may have been the only object to hold a reference to the
TestURLRequestContextGetter.  In the current codebase, however, my debugger
tells me there are two other outstanding references to it, so clearing the
profile's reference is not enough to ensure the object is deleted.  I'm hoping
this fix will be less likely to break as the code changes.  

This fixes issue 79933, but not the suppressions named after it.  One of those
suppressions seems to share a root cause with issue 80654, so it has been
renamed accordingly.  The other is so broad that it's not safe to delete it at
this time, because it's likely to be suppressing other bugs, too.

BUG=79933, 80654
TEST=MalwareDetailsTest.HTTPCache


Review URL: http://codereview.chromium.org/7919023

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 505d5b07b369d684c811f1347c65a6e098b2ec88