Fix InvalidationNotifierTest leak and suppressions
When XmppConnection is deleted, its TaskPump member is not deleted immediately.
The destructor enqueues a task to the message loop to have it deleted later.  A
comment in ~XmppConnection that explains why it does this.

In the unit tests, the TaskPump would get leaked because the message loop would
never get a chance to run that deletion task.  This commit fixes the problem by
running all pending tasks on that message loop before we exit the test.

This commit removes the suppression for bug_80238.  Issue 80238 hasn't been
seen for a while, so it's probably safe to remove.  It's only relation to issue
79651 is that it happened in the same set of tests; the two issues probably had
different root causes.

This commit renames the suppression bug_79651, which is still necessary, but
isn't actually related to issue 79651.  When I tried to remove it, I ran into
the issue now known as issue 96904.  I've renamed the suppression to match the
issue it's actually suppressing.

Discussion of these changes can be found on the bug tracker, spread across the
pages for issues 79651, 80238 and 96904.

BUG=79651, 80238, 96904
TEST=InvalidationNotifierTest


Review URL: http://codereview.chromium.org/7919026

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0a3eb07694609c091d3cd80ba72830164c947e8a