Remove suppression that is for an intentional static leak.
The line that we're hitting is:
DEFINE_STATIC_LOCAL(CSSStyleApplyProperty, cssStyleApplyPropertyInstance, ());

Valgrind knows to ignore this since it's reachable at shutdown.  It's just
counting it because it's in the suppressions.txt file.  Once we remove it,
it will no longer be counted as a leak.

BUG=79377

Review URL: http://codereview.chromium.org/8226022

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 46a6eb9a94fe9b9ba8e2ac0e9edf03a07f290fa5