Update expectations to remove all PERF_IMPROVE from the perf waterfall.
Some ranges were small, so I substituted the revb for 105360, which is in
the future. I will re-run the script in a few hours when we reach that
revision and make sure the new range is good.
Review URL: http://codereview.chromium.org/8275010

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5c133ca9ef4abcca79040045541c88b4fbfa8d04