Fix cause of uninitialized memory read in ExtensionService unit_tests
In crrev.com/113111 I checked in code that reads a app state boolean from an 
ExtensionSyncData object which is uninitialized. This fixes the initialization
for the relevant constructor, and also limits reading that value to only when
the extension is an app.

BUG=106549
TEST=valgrind run of ExtensionService sync-related unit_tests should come back
clean

Review URL: http://codereview.chromium.org/8899010

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b10a4abe46387826a54dd4a04fd251da948e10c6