chromeos: Merge tools/cros.DEPS/DEPS into the main DEPS.
This is to simplify setup needed to work on Chrome for Chrome OS.
cros.git and system_api.git are Chrome OS only, but pretty small.
The sizes of the repositories are 1.3MB and 260KB, and the former
will be gone in the not distant future.

BUG=108090
TEST=the error message is shown while gclient sync as intended if .gclient contains the deprecated cros_deps rule.


Review URL: http://codereview.chromium.org/9131003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 16ae48403450f92836bf216e7782b74c827f4b66