Revert 116766 - chromeos: Merge tools/cros.DEPS/DEPS into the main DEPS.
Reason: The new git workflow system failed to convert DEPS to .DEPS.git with this patch...

This is to simplify setup needed to work on Chrome for Chrome OS.
cros.git and system_api.git are Chrome OS only, but pretty small.
The sizes of the repositories are 1.3MB and 260KB, and the former
will be gone in the not distant future.

BUG=108090
TEST=the error message is shown while gclient sync as intended if .gclient contains the deprecated cros_deps rule.


Review URL: http://codereview.chromium.org/9131003

TBR=satorux@chromium.org
Review URL: http://codereview.chromium.org/9122018

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a67c002335d09b4b9035114a5651804eca9fbf7e