Add perf expectations for TEXTREL counts on 32-bit Linux binaries
We want sheriffs to notice if any TEXTRELs are added.  A count above zero
is unacceptable on anything but libffmpegsumo.so (which one day should get
down to zero as well, but it's known to be nontrivial).

BUG= 109318
TEST= none

R=cmp@chromium.org

Review URL: http://codereview.chromium.org/9187031

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 71e3dfb5bcd61bc06172a90461872465d6a82086