tools/android/forwarder (as well as tools/android/common)
'forwarder' is a tool that can forward a TCP port listening on the device
to a port on the host. It works like 'adb forward' but in the reverse
direction. It's useful to run tests on the device that need to access TCP
services on the host.

Review URL: http://codereview.chromium.org/9359003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c65c6e3150af3dc37471c76b10d2b05cf7367aa3