Make extract_directories work even if no --root-dir was specified.
trace_inputs tries to reduce the number of listed dependencies. When all files
in a directory are accessed, it replaces all the individual Results.File
instance by the Results.Directory instance representing the directory itself. It
is useful when hundreds of files are inside a directory and all of them are read
by an unit test.

The search is exhaustive and it is useful to have it stop at a directory. The
root directory used to be determined automatically but it is now explicit and
optional.

NOTRY=true
R=cmp@chromium.org
BUG=
TEST=


Review URL: https://chromiumcodereview.appspot.com/10541011

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5f7a6426efa5e968fa60c36cdd5866a3bcdf6cf7