Add a formal list of chrome-specific directories that cannot be tracked for now.
This is a side-effect that we set up so that the compiler slaves do not do a checkout of the test data for now. This may change eventually but it's not needed for now so just putting them in the untracked list is fine.

R=cmp@chromium.org
NOTRY=true
BUG=
TEST=


Review URL: https://chromiumcodereview.appspot.com/10556046

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 521dbcdd7e4c4064c6bed7f75a2836a1401fe5bf