Make checking the existence of the license file optional in the license tool
This is useful in situations where a checkout excludes some DEPS directories.
In this case, a checkout may include the README.chromium file for a particular
third-party project, but not the license file to which it refers.

This change adds a new require_license_file argument to licenses.ParseDir(), with
a default value of true. When true, there is no change in behaviour and if the
license file is missing, an exception is raised. When false, if the license
file is missing, we simply set the value of the 'License File' field of the
returned metadata to None.

This change also makes use of the new feature in
webview_licenses.GetIncompatibleDirectories() and
webview_licenses.GenerateNoticeFile().

BUG=138921

Review URL: https://chromiumcodereview.appspot.com/10829355

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2bb8cf46239ff12dde77b42644820a82e38ea9e6