Move all trace_inputs and isolate tests and their data into tests/ subdirectory.
Previously, the directory listing was a tad busy and it was hard to see what was
a tool or a unit test. 'data' was a generic name and it was unclear that it was
test data.

TBR=csharp@chromium.org
NOTRY=true
BUG=


Review URL: https://chromiumcodereview.appspot.com/10915048

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b7ff29bf05768d8f905e432152c1dc857b4b6650