Fix the unit test to run in a clean profile dir and add a new test to check proper migration.
The test was not producing correct results on one of the bots because the pref
checked has a special migration path that might populate it even if not set in the
policy file. To avoid this we run all tests from this group in a clean profile
directory and also to make sure migration works as expected I added a new test
to the suit too.

BUG=148356
TEST=unit_tests: DeviceSettingsProviderTest.SetPrefFailed should pass on all bots.


Review URL: https://chromiumcodereview.appspot.com/10916288

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fd8c8b580022c46e335709047f845d592d3cf850