Do not add <(PRODUCT_DIR)/First Run as a dependency, ever.
It is not generated by the build tool but a side-effect of browser_tests.
This should be fixed eventually; no test should ever write in PRODUCT_DIR, but
that's outside the scope of this effort.

R=csharp@chromium.org
NOTRY=true
BUG=


Review URL: https://chromiumcodereview.appspot.com/10984073

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9814c25674a412575b8c35eb77e3406f599d2a91