Scroll events used in the scrolling benchmarks should be located within the viewport
LayerTreeHostImpl::scrollBegin() is incorrectly returing ScrollIgnored
when running the scrolling benchmarks on sites having a scrollable area 
hight at least twice as big as the viewport height (consider 
white.bikeshed.org as an example). This is due to the scroll event
being located at the center of the scrollable area's bounds, and
thus located outside of the viewport's bounds.

This patch sets the location of scroll events to be centered in the
intersection of the scrollable area and the viewport. I am making
the assumption that the scrollable area and the viewport will
always have a non-empty intersection once the page loads, before any
scrolling takes place.

Running the benchmarks with the scrollcount.json page set (using --dont-override-profile) should report a non-0 value for numImplThreadScrolls.

BUG=


Review URL: https://chromiumcodereview.appspot.com/11380003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 061ea6b18c2513a8b16db30d57621dee781fd069