Increase timeout in Forwarder2's daemon.cc.
This is the timeout used to connect to the daemon's Unix Domain Socket. In case
the daemon was just spawned it can easily take a few hundred milliseconds on
the device before it binds the socket.

The same timeout is now also used when killing the daemon.


Review URL: https://chromiumcodereview.appspot.com/11418151

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5f0e88b4a552c9fb5b953c4495b652d0559be977