Telemetry: error messages & memory benchmark fix
Display better errors when clicking an element fails.

Also fixes the histogram printing for the memory benchmark. For a
hard-coded set of Google pages (such as GMail), we produce custom
memory histograms (V8.Something_gmail) instead of the generic
histograms (V8.Something), if we detect that a renderer is only
rendering this page and no other pages. For this test, we need to
disable histogram customizing, so that we get the same generic
histograms produced for all pages.

BUG=158323
NOTRY=true

Review URL: https://codereview.chromium.org/11478048

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c98594f646d3bd557c9441a2763cd1e263316491