[Telemetry] Fix an incorrect assertion.
We'll get an _OnClose during recording when ctrl+c'ing or when any other
exception is raised. In the first case, this just causes noise, but in the
second case it obscures the real cause of the failure.

BUG=171719

Review URL: https://codereview.chromium.org/11896077

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2f85d4cff273b56dd44d279be33a79a4e166e108