[telemetry] Recover gracefully when a tab crashes during tracing
Without this, the browser goes away, and _EndTracing explodes for lack of browser. Then, when that is fixed, the tracing thread's self._socket.recv() explodes because the self._socket was set to None when the browser was closed.

NOTRY=True

Review URL: https://chromiumcodereview.appspot.com/12095007

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5ac7ccd8bd7de550930601036d5dbd3ea7b09037